simpler design without NodeDelegate abstraction

+++ src/map_factory.cpp
@@ -9,7 +9,7 @@ namespace bpftrace {
 
 int MapFactory::call_count_ = 0;
 
-std::unique_ptr<IMap> MapFactory::constructMapForStoringBigStringsOffStack(int value_size) {
+std::unique_ptr<IMap> MapFactory::constructMapForStoringBigStringsOffStack(uint64_t value_size) {
 	std::string name = "str_call"+MapFactory::call_count_++;
 	return std::static_pointer_cast<IMap>(
 		std::make_unique<Map>(
